Avoid forward declaring the empty protocol
Currently, declarations like "CKAction<>" result in an empty forward declaration which doesn't compile ("@protocol ;"). This diff changes the behavior in the presence of the empty protocol to be to not forward declare. This is still kind of a bandaid solution for handling C++ classes in Remodel, full support for C++ is a more involved feature.