Add support for macros
Summary:
Currently we use static functions for nullability assertions, this leads to the assertion message being not as clear as it could be, as `__PRETTY_FUNCTION__` is just `RMParameterAssert`.

To improve this I will move `RMParameterAssert` to be a macro, so that the original assertion statement expands into the implementation. However, Remodel doesn't have support for generating macros currently. Therefore, I'm adding it. To maintain compatibility with existing plugins, I've made the `macros` field optional.

Reviewed By: calimarkus

Differential Revision: D8821966

fbshipit-source-id: f1be489ba8d513bcb68365ad4c33be830c684158