Add nullable annotation to ADT matchers
Summary:
ADT matchers perform `nil`-checking of the blocks that they're passed, but they don't have the `nullable` annotation. Allowing this to permit this to pass the compiler when `RMAssumeNonnull` is included.

For now I'm just doing `void` matchers, allowing `nil` for the `double`, `bool`, etc. cases is currently legal but defaults to `0`, so it's sort of a behavior change to permit it. That's probably fine, but let's split it up as this is unambigiously fine.

Reviewed By: calimarkus

Differential Revision: D13026965

fbshipit-source-id: 4fecefe4897eb7a6aee61d1401dcf3b585b5d8c1