Add command-line parameter to output into one file.
Summary:
This is part two of my endeavor to output everything into one source file in the hopes of making the buck experience better.

I've introduced a new plugin method which receives the current base file and the plugin can add stuff to it and return it. When the single file output flag is specified, this method is called instead of additionalFiles. This method could be extended later to be used for more thigs, e.g, setting file type. Instead of a separate file type method, we can just allow the plugin to modify the file type of the base file. I think this is the general direction that plugins should move it, so this is a good proof-of-concept to start with.

Reviewed By: ColinCampbell

Differential Revision: D14204157

fbshipit-source-id: af5f2ce00111fca840cf65c6b2a55fa5930c8dfd