Switch FileType to enum
Summary:
We currently use `===` to compare these when determining whether or not there's a conflict. `===` on objects compares identity, and since we create a new object every time we create one, this will never be `true`. Thus, if any two plugins specify a file type, even if they are the same, Remodel will error out. We should just use a TS enum instead, it's shorter anyways.

We also want to return an `Either`, not `throw` it.

Reviewed By: FieryRobot

Differential Revision: D15165757

fbshipit-source-id: 249aecbc3f77c2deed9de2e066b52093b08857af