Get additionalTypes such as fetch status to work with single file output
Summary: This allows plugins which return additionalTypes to now also be output in the same file for our single-file output option. It merges in the information that would have been output to a file into the base file. We may in the very near future deprecate additionalTypes altogether, but this gets things behaving more consistently for now.

Reviewed By: ColinCampbell

Differential Revision: D15236470

fbshipit-source-id: 683def3596f93289028d70a58bec198d2c8275df