Don't try to cast blocks to NSCopying
Summary:
The copying type safety plugin ensures that we don't generate an object that will immediately crash when initialized, because `-copy` is implemented on `NSObject`, but will crash for any object that doesn't conform to `NSCopying` (or implement `-copyWithZone:` without conforming). For better or worse, Remodel does have native support for `dispatch_block_t`, but if you write a `.value` file that declares an object of that type, it will fail on this plugin, because:

- It's not a `UIImage`, which does respond to `-copy`, but doesn't declare conformance.
- We should copy it.

So, we generate a line in the validator function that checks if it conforms to `NSCopying` at compile-time. However, `dispatch_block_t` doesn't conform to that protocol, even though it supports `-copy`. So, I am adding an additional hard-coded exception for `dispatch_block_t`.

Reviewed By: calimarkus

Differential Revision: D16072594

fbshipit-source-id: eb4de8cdab0412aaed9d970c3d1172260a2ed85d