NSLayoutAttribute is not part of Foundation
Summary:
Remodel has some code that skips emitting imports (for `%type`) if it thinks the type is in Foundation. But, it's wrong about `NSLayoutAttribute`: that's in UIKit!

https://developer.apple.com/documentation/uikit/nslayoutconstraint

I think we might be able to delete this check, we seem to deduplicate imports anyways, and double `#import` isn't harmful even if we don't. But I'm leaving that as a TODO in case I'm wrong.

(spent a while trying to figure out why the `%type` did not work for one specific type)

Reviewed By: ColinCampbell

Differential Revision: D22436008

fbshipit-source-id: b3a5f48c8ed091d4800e8b6aed2c1df45b761898