Add NS_SWIFT_NAME to Remodel ADT initializer methods
Summary:
Remodel ADTs are currently initialized in Swift like this:

```
# Example.adtValue
Example {
    Subtype {
        NSInteger integer
        NSString *string
    }
}

# Callsite.swift
let example = Example.subtype(with: 1, string: "test")
```

This is using Swift's native renaming, which takes only the `with`, not the rest of the selector keyword. For ADTs, it makes more sense to use the rest, but not the `with`. This makes them similar to how you'd write an initializer for a Swift `struct`, and how value objects are bridged:

```
# Callsite.swift
let example = Example.subtype(integer: 1, string: "test")
```

We can achieve this by using the `NS_SWIFT_NAME` attribute. We only need it for collection subtypes, single attribute subtypes are already okay as-is.

Reviewed By: manicakes

Differential Revision: D22947106

fbshipit-source-id: 37a5847540a9da36aa8adaa9ebdd61fad3f3d81c