Add fallbackDecoding keyword to enable renaming NSCoding value objects to ADTs
Summary:
As of now, there is no way to convert a .value object to an .adt if the object is being serialized and deserialized. The `initWithCoder:` method will not work, because it checks for a adt enum type first.

To fix this, we are adding a new keyword that will add a call to a helper function in the `initWithCoder:` initializer. This feeds in the NSCoder and the user can deserialize as needed.

See the test usage.

Reviewed By: natestedman

Differential Revision: D31090349

fbshipit-source-id: f16df06da6f6455c27913719c02915b6b4141826